+++ src/Nix/Effects.hs
@@ -448,10 +448,3 @@ addPath p =
 
 toFile_ :: (Framed e m, MonadStore m) => FilePath -> String -> m StorePath
 toFile_ p contents = addTextToStore (toText p) (toText contents) HS.empty False
-
-
--- * misc
-
--- Please, get rid of pathExists in favour of @doesPathExist@
-pathExists :: MonadFile m => FilePath -> m Bool
-pathExists = doesPathExist